# docs/api/openapi.v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "operationId": "OperationalController_plans_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Lists organization plans",
        "tags": [
          "Operational CRUD"
        ]
      },
      "post": {
        "operationId": "OperationalController_createPlan_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000000000
                  },
                  "chargeOpenDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31,
                    "default": 1
                  },
                  "chargeOpenTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                    "default": "00:00"
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "MONTHLY"
                    ],
                    "default": "MONTHLY"
                  }
                },
                "required": [
                  "name",
                  "amountCents",
                  "dueDay"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates an organization plan",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/plans/{id}": {
      "get": {
        "operationId": "OperationalController_plan_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets an organization plan",
        "tags": [
          "Operational CRUD"
        ]
      },
      "patch": {
        "operationId": "OperationalController_updatePlan_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000000000
                  },
                  "chargeOpenDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31,
                    "default": 1
                  },
                  "chargeOpenTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
                    "default": "00:00"
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "frequency": {
                    "type": "string",
                    "enum": [
                      "MONTHLY"
                    ],
                    "default": "MONTHLY"
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an organization plan",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/students": {
      "get": {
        "operationId": "OperationalController_students_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Lists organization students",
        "tags": [
          "Operational CRUD"
        ]
      },
      "post": {
        "operationId": "OperationalController_createStudent_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "cpf": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 14,
                    "description": "Brazilian CPF"
                  },
                  "rg": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30,
                    "description": "Brazilian RG"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 32
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates an organization student",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "operationId": "OperationalController_student_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets an organization student",
        "tags": [
          "Operational CRUD"
        ]
      },
      "patch": {
        "operationId": "OperationalController_updateStudent_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "cpf"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "cpf": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 14,
                    "description": "Brazilian CPF"
                  },
                  "rg": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 30,
                    "description": "Brazilian RG"
                  },
                  "birthDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "phone": {
                    "type": "string",
                    "maxLength": 32
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ]
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an organization student",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/students/{studentId}/guardians/{guardianId}": {
      "post": {
        "operationId": "OperationalController_link_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "relationship"
                ],
                "properties": {
                  "relationship": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Links a guardian to a student",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/guardians": {
      "get": {
        "operationId": "OperationalController_guardians_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Lists organization guardians",
        "tags": [
          "Operational CRUD"
        ]
      },
      "post": {
        "operationId": "OperationalController_createGuardian_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phone"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 32
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "taxId": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 14,
                    "description": "Brazilian CPF"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates an organization guardian",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/guardians/{id}": {
      "get": {
        "operationId": "OperationalController_guardian_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets an organization guardian",
        "tags": [
          "Operational CRUD"
        ]
      },
      "patch": {
        "operationId": "OperationalController_updateGuardian_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "phone",
                  "taxId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 32
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 255
                  },
                  "taxId": {
                    "type": "string",
                    "minLength": 11,
                    "maxLength": 14,
                    "description": "Brazilian CPF"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ]
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an organization guardian",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/enrollments": {
      "get": {
        "operationId": "OperationalController_enrollments_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Lists organization enrollments",
        "tags": [
          "Operational CRUD"
        ]
      },
      "post": {
        "operationId": "OperationalController_createEnrollment_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "studentId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "guardianId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "planId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000000000
                  },
                  "chargeOpenDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "chargeOpenTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "discountCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2000000000,
                    "default": 0
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "required": [
                  "studentId",
                  "guardianId",
                  "planId",
                  "startDate"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates an organization enrollment",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/enrollments/manual": {
      "post": {
        "operationId": "OperationalController_createStudentEnrollment_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Creates student, guardian, link and enrollment atomically",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
    "/api/v1/enrollments/{id}": {
      "get": {
        "operationId": "OperationalController_enrollment_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Gets an organization enrollment",
        "tags": [
          "Operational CRUD"
        ]
      },
      "patch": {
        "operationId": "OperationalController_updateEnrollment_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amountCents": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000000000
                  },
                  "chargeOpenDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "chargeOpenTime": {
                    "type": "string",
                    "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
                  },
                  "dueDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "discountCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2000000000,
                    "default": 0
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ENDED",
                      "CANCELLED"
                    ]
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Updates an organization enrollment",
        "tags": [
          "Operational CRUD"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mensaly API",
    "description": "Versioned backend API for Mensaly company owners and platform administrators",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mensaly_session"
      }
    },
...
  }
}